+++ rake.d/target.rb
@@ -157,7 +157,6 @@ class Target
     @libraries += list.collect do |entry|
       case entry
       when nil               then nil
-      when :magic            then c(:MAGIC_LIBS)
       when :flac             then c(:FLAC_LIBS)
       when :iconv            then c(:ICONV_LIBS)
       when :intl             then c(:LIBINTL_LIBS)